pyproject.toml
[tool.black]
line-length = 79

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.poetry]
name = "imgtools"
version = "0.1.0"
description = "Collection of tools that work on images"
authors = ["Harish Rajagopal <harish.rajagopals@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.19.5"
Pillow = "^8.1.0"
tqdm = "^4.56.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
